First re-factor to accomodate UNFCCC
