+++ tests/core/ingestion/test_validate_row.py
@@ -1,6 +1,6 @@
 from app.core.ingestion.ingest_row import DocumentIngestRow
 from app.core.ingestion.utils import IngestContext, ResultType
-from app.core.ingestion.validator import validate_document_row
+from app.core.ingestion.validator import validate_cclw_document_row
 from app.core.organisation import get_organisation_taxonomy
 
 from tests.core.ingestion.helpers import (
@@ -16,7 +16,7 @@ def test_validate_row__fails_bad_geography_iso(test_db):
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
     row.geography_iso = "XXX"
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -31,7 +31,7 @@ def test_validate_row__fails_empty_geography_iso(test_db):
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
     row.geography_iso = ""
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -45,9 +45,9 @@ def test_validate_row__consistent_family_and_collection(test_db):
     _, taxonomy = get_organisation_taxonomy(test_db, context.org_id)
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
     context.results = []
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -60,10 +60,10 @@ def test_validate_row__family_name_change(test_db):
     _, taxonomy = get_organisation_taxonomy(test_db, context.org_id)
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
     context.results = []
     row.family_name = "changed"
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -78,10 +78,10 @@ def test_validate_row__family_summary_change(test_db):
     _, taxonomy = get_organisation_taxonomy(test_db, context.org_id)
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
     context.results = []
     row.family_summary = "changed"
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -96,10 +96,10 @@ def test_validate_row__collection_name_change(test_db):
     _, taxonomy = get_organisation_taxonomy(test_db, context.org_id)
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
     context.results = []
     row.collection_name = "changed"
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -114,10 +114,10 @@ def test_validate_row__collection_summary_change(test_db):
     _, taxonomy = get_organisation_taxonomy(test_db, context.org_id)
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
     context.results = []
     row.collection_summary = "changed"
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -132,7 +132,7 @@ def test_validate_row__good_data(test_db):
     _, taxonomy = get_organisation_taxonomy(test_db, context.org_id)
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -146,7 +146,7 @@ def test_validate_row__bad_data(test_db):
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
     row.sectors = ["fish"]
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -160,7 +160,7 @@ def test_validate_row__resolvable_data(test_db):
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
     row.sectors = ["TranSPORtation"]
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -174,7 +174,7 @@ def test_validate_row__bad_document_type(test_db):
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
     row.document_type = "fish"
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -188,7 +188,7 @@ def test_validate_row__good_document_type(test_db):
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
     row.document_type = "Order"
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -202,7 +202,7 @@ def test_validate_row__bad_document_role(test_db):
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
     row.document_role = "fish"
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -216,7 +216,7 @@ def test_validate_row__good_document_role(test_db):
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
     row.document_role = "MAIN"
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -230,7 +230,7 @@ def test_validate_row__bad_document_variant(test_db):
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
     row.document_variant = "fish"
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
@@ -244,7 +244,7 @@ def test_validate_row__good_document_variant(test_db):
     row = DocumentIngestRow.from_row(1, get_doc_ingest_row_data(0))
     row.document_variant = "Translation"
 
-    validate_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
+    validate_cclw_document_row(test_db, context=context, row=row, taxonomy=taxonomy)
 
     assert context.results
     assert len(context.results) == 1
