+++ tests/routes/test_admin_unfccc.py
@@ -33,16 +33,16 @@ TWO_EVENT_ROWS = """Id,Eventable type,Eventable Id,Eventable name,Event type,Tit
 """
 
 
-def test_validate_bulk_ingest_cclw_law_policy(
+def test_validate_bulk_ingest_unfccc_law_policy(
     client,
     superuser_token_headers,
     test_db,
 ):
-    populate_taxonomy(db=test_db)
+    populate_taxonomy(test_db)
     populate_geography(test_db)
-    populate_document_type(db=test_db)
-    populate_document_role(db=test_db)
-    populate_document_variant(db=test_db)
+    populate_document_type(test_db)
+    populate_document_role(test_db)
+    populate_document_variant(test_db)
     test_db.commit()
     law_policy_csv_file = BytesIO(ONE_DFC_ROW.encode("utf8"))
     files = {
@@ -54,7 +54,7 @@ def test_validate_bulk_ingest_cclw_law_policy(
         ),
     }
     response = client.post(
-        "/api/v1/admin/bulk-ingest/validate/cclw/law-policy",
+        "/api/v1/admin/bulk-ingest/validate/unfccc/law-policy",
         files=files,
         headers=superuser_token_headers,
     )
@@ -67,7 +67,7 @@ def test_validate_bulk_ingest_cclw_law_policy(
     assert len(response_json["errors"]) == 0
 
 
-def test_bulk_ingest_cclw_law_policy(
+def test_bulk_ingest_unfccc_law_policy(
     client,
     superuser_token_headers,
     test_db,
@@ -76,11 +76,11 @@ def test_bulk_ingest_cclw_law_policy(
     mock_start_import = mocker.patch("app.api.api_v1.routers.admin._start_ingest")
     mock_write_csv_to_s3 = mocker.patch("app.api.api_v1.routers.admin.write_csv_to_s3")
 
-    populate_geography(db=test_db)
-    populate_taxonomy(db=test_db)
-    populate_document_type(db=test_db)
-    populate_document_role(db=test_db)
-    populate_document_variant(db=test_db)
+    populate_geography(test_db)
+    populate_taxonomy(test_db)
+    populate_document_type(test_db)
+    populate_document_role(test_db)
+    populate_document_variant(test_db)
     test_db.commit()
 
     law_policy_csv_file = BytesIO(ONE_DFC_ROW.encode("utf8"))
@@ -100,7 +100,7 @@ def test_bulk_ingest_cclw_law_policy(
         ),
     }
     response = client.post(
-        "/api/v1/admin/bulk-ingest/cclw/law-policy",
+        "/api/v1/admin/bulk-ingest/unfccc/law-policy",
         files=files,
         headers=superuser_token_headers,
     )
